Fix/ui fixes

* fixed school and student ui changes

* fixed ui

* fixed arrows

* Merge branch 'development' of https://bitbucket.org/aariya-tech/linguest-mobile-app-ui into fix/ui-fixes

* fixed bugs

* fixed create bot

* Add condition for Add Plan button.

* fixed bugs

* Merge branch development of https://bitbucket.org/aariya-tech/linguest-mobile-app-ui into fix/ui-fixes

* Merge branch 'temp' of https://bitbucket.org/aariya-tech/linguest-mobile-app-ui into fix/ui-fixes

* 1. Fix overlapping of iconWrapper.
2. Use dynamic color codes & remove unnecessary styling.

* Merge branch 'development' of https://bitbucket.org/aariya-tech/linguest-mobile-app-ui into fix/ui-fixes

* fixed signup/login flow, async storage, create bot and other ui/bug fixes

* added new cahtbot version

* Merge branch 'development' of https://bitbucket.org/aariya-tech/linguest-mobile-app-ui into fix/ui-fixes

* added image picker in the profile

* fixed redux's  states in some screens, fixed translations and other ui issues

* Merge branch 'development' of https://bitbucket.org/aariya-tech/linguest-mobile-app-ui into fix/ui-fixes

* fixed back arrow, text label in favorites, profile dob issue and other ui changes

* fixed notification,remove alert,delete icon,arrow back in drak mode

* Merge branch 'fix/ui-fixes' of https://bitbucket.org/aariya-tech/linguest-mobile-app-ui into fix/ui-fixes

* Merge branch 'development' of https://bitbucket.org/aariya-tech/linguest-mobile-app-ui into fix/ui-fixes

* alert on remove fav

* Merge branch 'development' of bitbucket.org:aariya-tech/linguest-mobile-app-ui into fix/ui-fixes


Approved-by: Vimit A

+++ components/Course.tsx
@@ -6,16 +6,16 @@ import { removeItemFromCart } from "@/redux/cartSlice";
 import { removeFromFavorite } from "@/redux/favoriteSlice";
 import { setLoader } from "@/redux/globalSlice";
 import { RootState } from "@/redux/store";
-import { FontAwesome, MaterialIcons } from "@expo/vector-icons";
 import { useRouter } from "expo-router";
 import React from "react";
 import { useTranslation } from "react-i18next";
-import { StyleSheet, Text, TouchableOpacity, ViewStyle } from "react-native";
+import { Alert, StyleSheet, Text, TouchableOpacity, ViewStyle } from "react-native";
 import { useDispatch, useSelector } from "react-redux";
 import AUIImage from "./common/AUIImage";
 import { AUIThemedText } from "./common/AUIThemedText";
 import { AUIThemedView } from "./common/AUIThemedView";
 import { ApiErrorToast, ApiSuccessToast } from "./common/AUIToast";
+import { MaterialCommunityIcons,MaterialIcons, FontAwesome } from '@expo/vector-icons';
 
 interface CourseProps {
     title: string;
@@ -72,16 +72,35 @@ const Course: React.FC<CourseProps> = ({
     });
 
     const handleRemoveFav = (id: string, type: string) => {
-        del(API_URL.favorite, { id: id, type: type })
-            .then((res: any) => {
-                ApiSuccessToast(res.message);
-                dispatch(removeFromFavorite({ id, type: "courses" }));
-                dispatch(setLoader(false));
-            })
-            .catch((e: any) => {
-                ApiErrorToast(e.response?.data?.message);
-                console.log(e);
-            });
+        Alert.alert(
+            'Remove Favorite',
+            `Are you sure you want to remove ${title} from your favorites?`,
+            [
+                {
+                    text: 'Cancel',
+                    onPress: () => console.log('Remove cancelled'),
+                    style: 'cancel',
+                },
+                {
+                    text: 'Remove',
+                    onPress: () => {
+                        dispatch(setLoader(true));
+                        del(API_URL.favorite, { id, type })
+                            .then((res: any) => {
+                                ApiSuccessToast(res.message);
+                                dispatch(removeFromFavorite({ id, type: "courses" }));
+                            })
+                            .catch((e: any) => {
+                                ApiErrorToast(e.response?.data?.message);
+                                console.log(e);
+                            })
+                            .finally(() => dispatch(setLoader(false)));
+                    },
+                    style: 'destructive',
+                },
+            ],
+            { cancelable: false }
+        );
     };
 
     return (
@@ -123,7 +142,7 @@ const Course: React.FC<CourseProps> = ({
                         onPress={() => handleRemoveFav(courseId, "course")}
                         style={styles.iconContainer}
                     >
-                        <MaterialIcons name="favorite" size={18} color="red" style={styles.icon} />
+                        <MaterialCommunityIcons name="delete-forever-outline" size={24} color="red" />
                     </TouchableOpacity>
                 )}
                 {edit && (
@@ -192,7 +211,7 @@ const styles = StyleSheet.create({
         position: "absolute",
         top: 10,
         right: 10,
-        backgroundColor: "rgba(255, 0, 0, 0.2)",
+        // backgroundColor: "rgba(255, 0, 0, 0.2)",
         borderRadius: 20,
         padding: 5,
     },
